Move tests for First and Combine into their own module
