+++ tests/test_cases/test_cocotb/Makefile
@@ -34,6 +34,7 @@ MODULE := "\
 	test_deprecated,\
 	test_doctests,\
 	test_synchronization_primitives,\
+	test_concurrency_primitives,\
 	"
 
 ifeq ($(shell python -c "import sys; print(sys.version_info >= (3, 6))"), "True")
