+++ tests/test_cases/test_cocotb/test_cocotb.py
@@ -891,40 +891,6 @@ def test_exceptions_forked(dut):
 
 
 @cocotb.test()
-def test_exceptions_first(dut):
-    """ Test exception propagation via cocotb.triggers.First """
-    @cocotb.coroutine
-    def raise_inner():
-        yield Timer(10)
-        raise ValueError('It is soon now')
-
-    @cocotb.coroutine
-    def raise_soon():
-        yield Timer(1)
-        yield cocotb.triggers.First(raise_inner())
-
-    # it's ok to change this value if the traceback changes - just make sure
-    # that when changed, it doesn't become harder to read.
-    expected = textwrap.dedent(r"""
-    Traceback \(most recent call last\):
-      File ".*test_cocotb\.py", line \d+, in _check_traceback
-        yield running_coro
-      File ".*test_cocotb\.py", line \d+, in raise_soon
-        yield cocotb\.triggers\.First\(raise_inner\(\)\)
-      File ".*triggers\.py", line \d+, in _wait
-        return await first_trigger[^\n]*
-      File ".*triggers.py", line \d+, in __await__
-        return \(yield self\)
-      File ".*triggers.py", line \d+, in __await__
-        return \(yield self\)
-      File ".*test_cocotb\.py", line \d+, in raise_inner
-        raise ValueError\('It is soon now'\)
-    ValueError: It is soon now""").strip()
-
-    yield _check_traceback(raise_soon(), ValueError, expected)
-
-
-@cocotb.test()
 def test_stack_overflow(dut):
     """
     Test against stack overflows when starting many coroutines that terminate
@@ -973,20 +939,6 @@ def test_immediate_coro(dut):
 
 
 @cocotb.test()
-def test_combine(dut):
-    """ Test the Combine trigger. """
-    # gh-852
-
-    @cocotb.coroutine
-    def do_something(delay):
-        yield Timer(delay)
-
-    crs = [cocotb.fork(do_something(dly)) for dly in [10, 30, 20]]
-
-    yield Combine(*(cr.join() for cr in crs))
-
-
-@cocotb.test()
 def test_clock_cycles_forked(dut):
     """ Test that ClockCycles can be used in forked coroutines """
     # gh-520
@@ -1001,96 +953,6 @@ def test_clock_cycles_forked(dut):
     b = cocotb.fork(wait_ten())
     yield a.join()
     yield b.join()
-
-
-@cocotb.test()
-def test_yield_list_stale(dut):
-    """ Test that a trigger yielded as part of a list can't cause a spurious wakeup """
-    # gh-843
-    events = [Event() for i in range(3)]
-
-    waiters = [e.wait() for e in events]
-
-    @cocotb.coroutine
-    def wait_for_lists():
-        ret_i = waiters.index((yield [waiters[0], waiters[1]]))
-        assert ret_i == 0, "Expected event 0 to fire, not {}".format(ret_i)
-
-        ret_i = waiters.index((yield [waiters[2]]))
-        assert ret_i == 2, "Expected event 2 to fire, not {}".format(ret_i)
-
-    @cocotb.coroutine
-    def wait_for_e1():
-        """ wait on the event that didn't wake `wait_for_lists` """
-        ret_i = waiters.index((yield waiters[1]))
-        assert ret_i == 1, "Expected event 1 to fire, not {}".format(ret_i)
-
-    @cocotb.coroutine
-    def fire_events():
-        """ fire the events in order """
-        for e in events:
-            yield Timer(1)
-            e.set()
-
-    fire_task = cocotb.fork(fire_events())
-    e1_task = cocotb.fork(wait_for_e1())
-    yield wait_for_lists()
-
-    # make sure the other tasks finish
-    yield fire_task.join()
-    yield e1_task.join()
-
-
-@cocotb.test()
-def test_nested_first(dut):
-    """ Test that nested First triggers behave as expected """
-    events = [Event() for i in range(3)]
-    waiters = [e.wait() for e in events]
-
-    @cocotb.coroutine
-    def fire_events():
-        """ fire the events in order """
-        for e in events:
-            yield Timer(1)
-            e.set()
-
-
-    @cocotb.coroutine
-    def wait_for_nested_first():
-        inner_first = First(waiters[0], waiters[1])
-        ret = yield First(inner_first, waiters[2])
-
-        # should unpack completely, rather than just by one level
-        assert ret is not inner_first
-        assert ret is waiters[0]
-
-    fire_task = cocotb.fork(fire_events())
-    yield wait_for_nested_first()
-    yield fire_task.join()
-
-
-@cocotb.test()
-async def test_first_does_not_kill(dut):
-    """ Test that `First` does not kill coroutines that did not finish first """
-    ran = False
-
-    # decorating `async def` is required to use `First`
-    @cocotb.coroutine
-    async def coro():
-        nonlocal ran
-        await Timer(2, units='ns')
-        ran = True
-
-    # Coroutine runs for 2ns, so we expect the timer to fire first
-    timer = Timer(1, units='ns')
-    t = await First(timer, coro())
-    assert t is timer
-    assert not ran
-
-    # the background routine is still running, but should finish after 1ns
-    await Timer(2, units='ns')
-
-    assert ran
 
 
 @cocotb.test()
